# apiDocumentation/swagger.yaml
openapi: 3.0.0
info:
  description: This is the ISE-Learning REST API
  version: 0.0.0
  title: ISE-Learning REST API
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  '/v0/users/{userId}':  
    get:
      tags:
        - Users
      summary: Retreives an user
      operationId: getUser
      description: |
        Retrive an user's information.
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: User's unique id.
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/user'
    delete:
      tags:
        - Users
      summary: Deletes an user
      operationId: deleteUser
      description: |
        Delete an user.
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: User's unique id.
      responses:
        '204':
          description: Successful deleted user.
          content:
            application/json:
              schema:
                type: object   
...
  '/v0/challenges/questionnaires/{questionnaireId}':
    get:
      tags:
        - Challenges
      summary: Retreives all challenges from a given questionnaire
      operationId: getAllChallengesByQuestionnaireId
      description: |
        Retrieves every challenge from a questionnaire by its unique identifier.
      parameters:
        - in: path
          name: questionnaireId
          schema:
            type: integer
          required: true
          description: Questionnaire unique id.
      responses:
        '200':
          description: Challenge
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/challenge'
...
  '/v0/questionnaires':
    post:
      tags:
        - Questionnaires
      summary: Creates a questionnaire
      operationId: createQuestionnaire
      description: |
        Creates a questionnaire based on payload content.
      requestBody:
        description: Questionnaire to be created.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/questionnaireModel'
      responses:
        '201':
          description: Successful questionnaire created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaire'
  '/v0/questionnaires/withChallenges':
    post:
      tags:
        - Questionnaires
      summary: Creates a questionnaire with challenges
      operationId: createQuestionnaireWithChallenges
      description: |
        Creates a questionnaire based on payload content.
      requestBody:
        description: Questionnaire to be created, and the challenges to associate to it.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/questionnaireWithChallengesModel'
      responses:
        '201':
          description: Successful questionnaire created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaire'
  '/v0/questionnaires/{questionnaireId}':
    get:
      tags:
        - Questionnaires
      summary: Retrieves a questionnaire
      operationId: getQuestionnaireById
      description: |
        Retrieves a single questionnaire by its unique identifer.
      parameters:
        - in: path
          name: questionnaireId
          schema:
            type: integer
          required: true
          description: Questionnaire unique identifer.
      responses:
        '200':
          description: Questionnaire
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaire'
    put:
      tags:
        - Questionnaires
      summary: Updates a questionnaire
      operationId: updateQuestionnaireById
      description: |
        Updates a questionnaire by its unique identifier.
      parameters:
        - in: path
          name: questionnaireId
          schema:
            type: integer
          required: true
          description: Questionnaire unique identifer.
      requestBody:
        description: Questionnaire to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/questionnaireModel'
      responses:
        '200':
          description: Successful updated questionnaire.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaire'
    delete:
      tags:
        - Questionnaires
      summary: Deletes a questionnaire
      operationId: deleteQuestionnaireById
      description: |
        Delete a questionnaire by its unique identifier.
      parameters:
        - in: path
          name: questionnaireId
          schema:
            type: integer
          required: true
          description: Questionnaire unique identifer.
      responses:
        '204':
          description: Successful deleted questionnaire.
          content:
            application/json:
              schema:
                type: object
  '/v0/questionnaires/users/{userId}':
    get:
      tags:
        - Questionnaires
      summary: Retrieves all questionnaires of specified user
      operationId: getAllUserQuestionnaires
      description: |
        Retrieves a collection of questionnaires by its creator unique identifer.
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: User unique identifer.
      responses:
        '200':
          description: User's questionnaire collection.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaire'
...
  '/v0/questionnaireInstances/questionnaires/{questionnaireId}':
    get:
      tags:
        - Questionnaire Instances
      summary: Retrieves a all of questionnaire instances from a questionnaire
      operationId: getAllQuestionnaireInstancesByQuestionnaireId
      description: |
        Retrieves a collection of questionnaire instances from a questionnaire by its universal unique identifer.
      parameters:
        - in: path
          name: questionnaireId
          schema:
            type: integer
          required: true
          description: Questionnaire unique identifer.
      responses:
        '200':
          description: Questionnaire's questionnaire instance collection.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaireInstance'
...
components:
  schemas:
...
    user:
      type: object
      properties:
        userId: 
          type: integer
          example: 523
          description: Unique id which uniquely identifies the user
        username:
          type: string
          example: "userOne"
          description: Unique string used to identify an user
        email:
          type: string
          example: "userOne@example.com"
          description: Email address which can be used to contact user
        name:
          type: string
          example: "John Doe"
          description: User's given name
...
    challenge:
      type: object
      required:
        - challengeID
        - challengeText
        - isPrivate
        - solutions
      properties:
        challengeID:
          type: integer
          example: 548
          description: Unique identifier of challenge.
        challengeText:
          type: string
          example: 'Implement Hello world, by printing Hello world into the console.'
          description: Test describing the challenge and its goal.
        isPrivate:
          type: boolean
          example: true
          description: >-
            Boolean value indicating whether the challenge can only be seen by
            its creator
        solutions:
          type: array
          description: List of challenges.
          items:
            $ref: '#/components/schemas/solution'
...
    solution:
      type: object
      required:
        - solutionId
        - codeLanguage
        - challengeCode
        - solutionCode
        - unitTests
      properties:
        solutionId:
          type: integer
          example: 79884
          description: Unique id to which uniquely identifies the solution.
        codeLanguage:
          type: string
          example: java
          description: Text specifying in which language the solution is.
        challengeCode:
          type: string
          example: 'public class code { public static void main(String[] args) { }}'
          description: Default code of the code presented to user.
        solutionCode:
          type: string
          example: >-
            public class code { public static void main(String[] args) {
            System.out.println(\"Hello World\");}}
          description: Solution of the challenge.
        unitTests:
          type: string
          example: >-
            import org.junit.Assert;import org.junit.Test;public class unitTests
            {@Test public void TestCase1() {Assert.assertTrue(true);}}
          description: Unit tests to which the the user submitted will be tested against.
...
    questionnaire:
      type: object
      required:
        - questionnaireId
        - description
        - timer
        - creatorId
        - creationDate
      properties:
        questionnaireId:
          type: integer
          example: 21313
          description: questionnaire unique identifier.
        description:
          type: string
          example: Questionnaire for course PS-1920SV
          description: Text describing the questionnaire.
        timer:
          type: number
          example: 360000
          description: >-
            Time in milliseconds which an user has to solve the questionnaire.
            If value is NULL has no time limite.
        creatorId:
          type: integer
          example: 44598
          description: Unique user identifier of user who created the questionnaire.
        creationDate:
          type: string
          example: '2020/05/12'
          description: questionnaire's creation date in yyyy-MM-dd format
    questionnaireModel:
      type: object
      required:
        - description
        - timer
        - creatorId
      properties:
        description:
          type: string
          example: Questionnaire for course PS-1920SV
          description: Text describing the questionnaire.
        timer:
          type: number
          example: 360000
          description: >-
            Time in milliseconds which an user has to solve the questionnaire.
            If value is NULL has no time limite.
        creatorId:
          type: integer
          example: 44598
          description: Unique user identifier of user who created the questionnaire.
    questionnaireWithChallengesModel:
      type: object
      required:
        - questionnaire
        - challenges
      properties:
        questionnaire:
          $ref: '#/components/schemas/questionnaireModel'
        challenges:
          type: array
          items:
            $ref: '#/components/schemas/questionnaireChallengeCollectionModel'
    questionnaireInstance:
      type: object
      required:
        - questionnaireId
        - questionnaireInstanceId
        - questionnaireInstanceUuid
        - description
        - timer
        - startTimestamp
        - endTimestamp
        - isFinished
      properties:
        questionnaireInstanceId:
          type: integer
          example: 85674
          description: Questionnaire instance unique identifier.
        questionnaireId:
          type: integer
          example: 2213
          description: Questionnaire unique identifier.
        questionnaireInstanceUuid:
          type: string
          example: 'c2c19e60-c32c-4413-84fc-369d6603d5be'
          description: Questionnaire universal unique identifier.
        description:
          type: string
          example: Questionnaire for course PS-1920SV and class LI61N 
          description: Text describing the questionnaire instance.
        timer:
          type: number
          example: 1800000
          description: >-
            Time in milliseconds which an user has to solve the questionnaire instance.
            If value is NULL has no time limite.
        startTimestamp:
          type: number
          example: 231231230000000
          description: >- 
            Time in milliseconds that user started solving the questionnaire instance.
        endTimestamp:
          type: number
          example: 231231231800000
          description: >-
            Time in milliseconds that user finished solving the questionnaire instance.
        isFinished:
          type: boolean
          example: false
          description: Indicates if user can still access questionnaire instance.
...
    questionnaireChallengeCollectionModel:
      type: object
      required:
        - challengeId
        - languageFilter
      properties:
        challengeId:
          type: integer
          example: 2342
          description: Unique identifier of challenge.
        languageFilter:
          type: string
          example: 'java,javascript,python'
          description: >-
            Programming languages allowed to be used to solve the questionnaire challenge, separeted by commas.
...